+++ backend/de.metas.dunning/src/main/java/de/metas/dunning/invoice/model/validator/C_Invoice.java
@@ -1,60 +1,30 @@
 package de.metas.dunning.invoice.model.validator;
 
-import static org.adempiere.model.InterfaceWrapperHelper.getTableId;
-
-/*
- * #%L
- * de.metas.dunning
- * %%
- * Copyright (C) 2015 metas GmbH
- * %%
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as
- * published by the Free Software Foundation, either version 2 of the
- * License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public
- * License along with this program.  If not, see
- * <http://www.gnu.org/licenses/gpl-2.0.html>.
- * #L%
- */
-
-
-import java.sql.Timestamp;
-import java.time.LocalDate;
-import java.util.List;
-import java.util.Properties;
-
-import org.adempiere.ad.modelvalidator.annotations.DocValidate;
-import org.adempiere.ad.modelvalidator.annotations.ModelChange;
-import org.adempiere.ad.modelvalidator.annotations.Validator;
-import org.adempiere.model.InterfaceWrapperHelper;
-import org.compiere.SpringContextHolder;
-import org.compiere.model.ModelValidator;
-import org.compiere.util.TimeUtil;
-import org.slf4j.Logger;
-
 import de.metas.adempiere.model.I_C_Invoice;
 import de.metas.dunning.api.IDunningBL;
 import de.metas.dunning.api.IDunningContext;
 import de.metas.dunning.api.IDunningDAO;
-import de.metas.invoice.due_date.InvoiceDueDateProviderService;
 import de.metas.dunning.invoice.api.IInvoiceSourceBL;
 import de.metas.dunning.model.I_C_Dunning_Candidate;
-import de.metas.invoice.InvoiceId;
 import de.metas.logging.LogManager;
 import de.metas.util.Services;
+import org.adempiere.ad.modelvalidator.annotations.DocValidate;
+import org.adempiere.ad.modelvalidator.annotations.ModelChange;
+import org.adempiere.ad.modelvalidator.annotations.Validator;
+import org.adempiere.model.InterfaceWrapperHelper;
+import org.compiere.model.ModelValidator;
+import org.slf4j.Logger;
+
+import java.sql.Timestamp;
+import java.util.List;
+import java.util.Properties;
+
+import static org.adempiere.model.InterfaceWrapperHelper.getTableId;
 
 @Validator(I_C_Invoice.class)
 public class C_Invoice
 {
 	private static final Logger logger = LogManager.getLogger(C_Invoice.class);
-	private final InvoiceDueDateProviderService invoiceDueDateProviderService = SpringContextHolder.instance.getBean(InvoiceDueDateProviderService.class);
 
 	@DocValidate(timings = { ModelValidator.TIMING_AFTER_PREPARE })
 	public void setDunningGraceIfAutomatic(final I_C_Invoice invoice)
@@ -64,25 +34,9 @@ public class C_Invoice
 	}
 
 	/**
-	 * This shall set the Due Date in invoice considering payment term or contracts, but only if due date was not set previously
-	 */
-	@DocValidate(timings = { ModelValidator.TIMING_AFTER_PREPARE })
-	public void setDueDate(final I_C_Invoice invoice)
-	{
-		if (invoice.getDueDate() == null)
-		{
-			final LocalDate dueDate = invoiceDueDateProviderService.provideDueDateFor(InvoiceId.ofRepoId(invoice.getC_Invoice_ID()));
-			invoice.setDueDate(TimeUtil.asTimestamp(dueDate));
-			InterfaceWrapperHelper.save(invoice);
-		}
-	}
-
-	/**
 	 * This method is triggered when DunningGrace field is changed.
 	 *
-	 * NOTE: to developer: please keep this method with only ifColumnsChanged=DunningGrace because we want to avoid update cycles between invoice and dunning candidate
-	 *
-	 * @param invoice
+	 * @implNote please keep this method with only ifColumnsChanged=DunningGrace because we want to avoid update cycles between invoice and dunning candidate
 	 */
 	@ModelChange(timings = ModelValidator.TYPE_AFTER_CHANGE
 			, ifColumnsChanged = I_C_Invoice.COLUMNNAME_DunningGrace)
